requirements.txt
fastapi
typer[all]
google-cloud-firestore
together
openai
pydantic
bandit
uvicorn
python-dotenv